# docs/openapi.yaml
openapi: 3.1.0
info:
  title: API del Observatorio de Semiconductores
  version: 0.5.6
  description: >-
    Contrato de autenticación, catálogos y las verticales de señales,
    tendencias, alertas y las APIs editorial interna y pública, incluida la
    resolución de bloques, entrega de medios y exportaciones autenticadas,
    implementadas hasta la Fase 5.6.
servers:
  - url: /api
...
paths:
...
  /catalogs/{catalog}:
    get:
      tags: [Catalogs]
      summary: Obtiene los elementos de un catálogo
      operationId: getCatalog
      parameters:
        - name: catalog
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/CatalogName"
      responses:
        "200":
          description: Catálogo obtenido
          content:
            application/json:
              schema:
                type: object
                required: [success, message, data]
                properties:
                  success: { type: boolean, const: true }
                  message: { type: string }
                  data:
                    type: object
                    required: [name, items]
                    properties:
                      name: { $ref: "#/components/schemas/CatalogName" }
                      items:
                        type: array
                        items:
                          type: object
                          additionalProperties: true
        "404":
          $ref: "#/components/responses/NotFoundError"
  /signals:
    get:
      tags: [Signals]
      summary: Lista exclusivamente señales validadas
      operationId: listPublicSignals
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/CategoryId"
        - $ref: "#/components/parameters/Fcv"
        - $ref: "#/components/parameters/Impact"
        - $ref: "#/components/parameters/Urgency"
        - $ref: "#/components/parameters/Reliability"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
        - $ref: "#/components/parameters/SignalSort"
      responses:
        "200":
          description: Señales públicas paginadas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedSignalsResponse" }
        "400":
          $ref: "#/components/responses/ValidationError"
  /signals/{id}:
    get:
      tags: [Signals]
      summary: Obtiene una señal validada
      operationId: getPublicSignal
      parameters:
        - $ref: "#/components/parameters/SignalId"
      responses:
        "200":
          description: Señal pública
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignalResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
  /admin/signals:
    get:
      tags: [Admin Signals]
      summary: Lista señales internas
      operationId: listInternalSignals
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/SignalStatusCode" }
        - $ref: "#/components/parameters/CategoryId"
        - $ref: "#/components/parameters/Fcv"
        - $ref: "#/components/parameters/Impact"
        - $ref: "#/components/parameters/Urgency"
        - $ref: "#/components/parameters/Reliability"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
        - $ref: "#/components/parameters/SignalSort"
      responses:
        "200":
          description: Señales internas paginadas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedSignalsResponse" }
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/PermissionError"
    post:
      tags: [Admin Signals]
      summary: Crea una señal NEW para el analista autenticado
      operationId: createSignal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSignalInput" }
      responses:
        "201":
          description: Señal creada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignalResponse" }
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/PermissionError"
        "422":
          $ref: "#/components/responses/DomainRuleError"
  /admin/signals/{id}:
    get:
      tags: [Admin Signals]
      summary: Obtiene una señal con campos internos
      operationId: getInternalSignal
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SignalId"
      responses:
        "200":
          description: Señal interna
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignalResponse" }
        "404":
          $ref: "#/components/responses/NotFoundError"
    patch:
      tags: [Admin Signals]
      summary: Modifica una señal en estado NEW
      operationId: updateSignal
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SignalId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSignalInput" }
      responses:
        "200":
          description: Señal actualizada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignalResponse" }
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/DomainRuleError"
  /admin/signals/{id}/transitions:
    post:
      tags: [Admin Signals]
      summary: Ejecuta una transición semántica de estado
      operationId: transitionSignal
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SignalId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignalTransitionInput" }
      responses:
        "200":
          description: Estado actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignalResponse" }
        "403":
          $ref: "#/components/responses/PermissionError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/DomainRuleError"
  /admin/signals/{id}/history:
    get:
      tags: [Admin Signals]
      summary: Obtiene el historial de estados de una señal
      operationId: getSignalHistory
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SignalId"
      responses:
        "200":
          description: Historial cronológico
          content:
            application/json:
              schema:
                type: object
                required: [success, message, data]
                properties:
                  success: { type: boolean, const: true }
                  message: { type: string }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/SignalHistoryEntry" }
...
  /admin/actors:
    get:
      tags: [Admin Trends]
      summary: Lista actores disponibles para tendencias
      operationId: listActors
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Search" }]
      responses:
        "200":
          description: Actores disponibles
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SignalId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Search:
      name: search
      in: query
      schema: { type: string, minLength: 2, maxLength: 200 }
    CategoryId:
      name: categoryId
      in: query
      schema: { type: integer, minimum: 1 }
    Fcv:
      name: fcv
      in: query
      schema: { type: string }
    Impact:
      name: impact
      in: query
      schema: { type: string, enum: [LOW, MEDIUM, HIGH] }
    Urgency:
      name: urgency
      in: query
      schema: { type: string, enum: [LOW, MEDIUM, HIGH] }
    Reliability:
      name: reliability
      in: query
      schema: { type: string, enum: [LOW, MEDIUM, HIGH] }
    Scope:
      name: scope
      in: query
      schema: { type: string }
    FromDate:
      name: from
      in: query
      schema: { type: string, format: date }
    ToDate:
      name: to
      in: query
      schema: { type: string, format: date }
    SignalSort:
      name: sort
      in: query
      schema:
        type: string
        default: -publicationDate
        enum:
          - publicationDate
          - -publicationDate
          - captureDate
          - -captureDate
          - ips
          - -ips
          - title
          - -title
          - updatedAt
          - -updatedAt
...
  schemas:
...
    CatalogName:
      type: string
      enum:
        - source-types
        - categories
        - fcv
        - block-types
        - signal-types
        - reliability-levels
        - impacts
        - urgencies
        - scopes
        - signal-statuses
        - trend-maturity
        - trend-statuses
        - trend-directions
        - alert-levels
        - alert-statuses
        - alert-origins
        - audiences
        - content-types
        - content-statuses
        - file-types
        - actor-types
        - content-relation-types
        - section-types
    SignalStatusCode:
      type: string
      enum: [NEW, UNDER_REVIEW, VALIDATED, ARCHIVED]
    CatalogValue:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
    Signal:
      type: object
      required:
        - id
        - businessCode
        - title
        - summary
        - publicationDate
        - captureDate
        - evidenceUrl
        - ips
        - priority
        - category
        - source
        - signalType
        - impact
        - urgency
        - reliability
        - scope
        - status
        - keywords
        - relations
        - createdAt
        - updatedAt
      properties:
        id: { type: integer, format: int64 }
        businessCode: { type: string, example: SIG-2026-000001 }
        title: { type: string }
        summary: { type: string }
        publicationDate: { type: string, format: date }
        captureDate: { type: string, format: date }
        evidenceUrl: { type: string, format: uri }
        ips: { type: number, minimum: 1, maximum: 27 }
        priority: { type: string, enum: [LOW, MEDIUM, HIGH] }
        category:
          type: object
          required: [id, name, fcv]
          properties:
            id: { type: integer }
            name: { type: string }
            fcv: { $ref: "#/components/schemas/CatalogValue" }
        source:
          type: object
          required: [id, name]
          properties:
            id: { type: integer, format: int64 }
            name: { type: string }
        signalType: { $ref: "#/components/schemas/CatalogValue" }
        impact: { $ref: "#/components/schemas/CatalogValue" }
        urgency: { $ref: "#/components/schemas/CatalogValue" }
        reliability: { $ref: "#/components/schemas/CatalogValue" }
        scope: { $ref: "#/components/schemas/CatalogValue" }
        status: { $ref: "#/components/schemas/CatalogValue" }
        keywords:
          type: array
          items:
            type: object
            required: [id, name]
            properties:
              id: { type: integer, format: int64 }
              name: { type: string }
        relations:
          type: object
          required: [linkedToTrend, linkedToAlert]
          properties:
            linkedToTrend: { type: boolean }
            linkedToAlert: { type: boolean }
        analyst:
          type: object
          description: Solo se incluye en respuestas internas.
          properties:
            id: { type: integer, format: int64 }
            name: { type: string }
        validator:
          type: [object, "null"]
          description: Solo se incluye en respuestas internas.
        validationDate: { type: [string, "null"], format: date }
        notes: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SignalResponse:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: "#/components/schemas/Signal" }
    Pagination:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        totalItems: { type: integer }
        totalPages: { type: integer }
    PaginatedSignalsResponse:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data:
          type: object
          required: [items, pagination]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Signal" }
            pagination: { $ref: "#/components/schemas/Pagination" }
    CreateSignalInput:
      type: object
      additionalProperties: false
      required:
        - title
        - summary
        - publicationDate
        - evidenceUrl
        - categoryId
        - sourceId
        - signalTypeCode
        - impactCode
        - urgencyCode
        - reliabilityCode
        - scopeCode
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        summary: { type: string, minLength: 1, maxLength: 5000 }
        publicationDate: { type: string, format: date }
        evidenceUrl: { type: string, format: uri, maxLength: 2000 }
        categoryId: { type: integer, minimum: 1 }
        sourceId: { type: integer, format: int64, minimum: 1 }
        signalTypeCode: { type: string }
        impactCode: { type: string, enum: [LOW, MEDIUM, HIGH] }
        urgencyCode: { type: string, enum: [LOW, MEDIUM, HIGH] }
        reliabilityCode: { type: string, enum: [LOW, MEDIUM, HIGH] }
        scopeCode: { type: string }
        notes: { type: [string, "null"], maxLength: 5000 }
        keywords:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 80 }
    UpdateSignalInput:
      type: object
      additionalProperties: false
      minProperties: 1
      description: Todos los campos son opcionales; IPS, estado y autor no se aceptan.
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        summary: { type: string, minLength: 1, maxLength: 5000 }
        publicationDate: { type: string, format: date }
        evidenceUrl: { type: string, format: uri, maxLength: 2000 }
        categoryId: { type: integer, minimum: 1 }
        sourceId: { type: integer, format: int64, minimum: 1 }
        signalTypeCode: { type: string }
        impactCode: { type: string, enum: [LOW, MEDIUM, HIGH] }
        urgencyCode: { type: string, enum: [LOW, MEDIUM, HIGH] }
        reliabilityCode: { type: string, enum: [LOW, MEDIUM, HIGH] }
        scopeCode: { type: string }
        notes: { type: [string, "null"], maxLength: 5000 }
        keywords:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 80 }
        updatedAt: { type: string, format: date-time }
    SignalTransitionInput:
      type: object
      additionalProperties: false
      required: [transition]
      properties:
        transition:
          type: string
          enum:
            - SUBMIT_FOR_REVIEW
            - REQUEST_CHANGES
            - VALIDATE
            - REOPEN
            - ARCHIVE
        notes: { type: [string, "null"], maxLength: 2000 }
    SignalHistoryEntry:
      type: object
      required: [id, toStatus, transition, changedBy, changedAt]
      properties:
        id: { type: integer, format: int64 }
        fromStatus:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/CatalogValue"
        toStatus: { $ref: "#/components/schemas/CatalogValue" }
        transition: { type: string }
        changedBy:
          type: object
          required: [id, name]
          properties:
            id: { type: integer, format: int64 }
            name: { type: string }
        notes: { type: [string, "null"] }
        changedAt: { type: string, format: date-time }
...
    ErrorDetail:
      type: object
      required: [field, message]
      properties:
        field: { type: [string, "null"] }
        message: { type: string }
    ErrorResponse:
      type: object
      required: [success, message, code, errors]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        code: { type: string }
        errors:
          oneOf:
            - type: "null"
            - type: array
              items: { $ref: "#/components/schemas/ErrorDetail" }
  responses:
...
    ValidationError:
      description: Datos inválidos
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    AuthenticationError:
      description: Autenticación requerida o inválida
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PermissionError:
      description: Permiso insuficiente
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFoundError:
      description: Recurso inexistente
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ConflictError:
      description: Conflicto con el estado actual
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    DomainRuleError:
      description: La operación incumple una regla de dominio
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }